Changed file structure, added points system
